requirements.txt
Jinja2==3.1.4
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.18.0
SQLAlchemy==2.0.36
annotated-types==0.7.0
anyio==4.7.0
audioread==3.0.1
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi-cli==0.0.7
fastapi==0.115.6
ffmpeg-python==0.2.0
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio_ffmpeg==0.5.1
itsdangerous==2.2.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
moviepy==2.1.1
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
orjson==3.10.12
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
psycopg2-binary==2.9.9
pycparser==2.22
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
rich-toolkit==0.12.0
rich==13.9.4
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.13.1
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
